fix minor changes

+++ get_prod-out.php
@@ -35,47 +35,53 @@ if(isset($_GET["action"]) && $_GET["action"] === "endorse_product") {
    $sql = "SELECT * FROM `$table` WHERE endorsed_by='$endorsed_by';";
    $result = mysqli_query($conn, $sql);
    
-   while ($row = mysqli_fetch_array($result)) {
-      $sql2 = "INSERT INTO endorse_final (
-         barcode, 
-         description, 
-         quantity, 
-         lot,
-         branch, 
-         mrf,
-         order_num, 
-         exp_date,
-         remarks, 
-         endorsed_by,
-         endorsed_date) 
-         VALUES (
-         '" . $row["barcode"] . "', 
-         '" . $row["description"] . "', 
-         '" . $row["quantity"] . "', 
-         '" . $row["lot"] . "',
-         '" . $row["branch"] . "', 
-         '" . $row["mrf"] . "',
-         '" . $row["order_num"] . "', 
-         '" . $row["exp_date"] . "',
-         '" . $row["remarks"] . "', 
-         '" . $row["endorsed_by"] . "',
-         '" . $row["endorsed_date"] . "')";
-
-      if (mysqli_query($conn, $sql2)) {
-         $data = 'success';
-         echo json_encode($data);
-      } else {
-         echo "Error inserting data: " . mysqli_error($conn);
+   if (!$result->num_rows < 0){
+      while ($row = mysqli_fetch_array($result)) {
+         $sql2 = "INSERT INTO endorse_final (
+            barcode, 
+            description, 
+            quantity, 
+            lot,
+            branch, 
+            mrf,
+            order_num, 
+            exp_date,
+            remarks, 
+            endorsed_by,
+            endorsed_date) 
+            VALUES (
+            '" . $row["barcode"] . "', 
+            '" . $row["description"] . "', 
+            '" . $row["quantity"] . "', 
+            '" . $row["lot"] . "',
+            '" . $row["branch"] . "', 
+            '" . $row["mrf"] . "',
+            '" . $row["order_num"] . "', 
+            '" . $row["exp_date"] . "',
+            '" . $row["remarks"] . "', 
+            '" . $row["endorsed_by"] . "',
+            '" . $row["endorsed_date"] . "')";
+   
+         if (mysqli_query($conn, $sql2)) {
+            $data = 'success';
+            echo json_encode($data);
+         } else {
+            echo "Error inserting data: " . mysqli_error($conn);
+         }
+   
+         $sql3 = "DELETE FROM endorse WHERE endorsed_by='$endorsed_by';";
+   
+         if (mysqli_query($conn, $sql3)) {
+            echo "Data deleted successfully";
+         } else {
+            echo "Error deleting data: " . mysqli_error($conn);
+         }
       }
-
-      $sql3 = "DELETE FROM endorse WHERE endorsed_by='$endorsed_by';";
-
-      if (mysqli_query($conn, $sql3)) {
-         echo "Data deleted successfully";
-      } else {
-         echo "Error deleting data: " . mysqli_error($conn);
-      }
+   } else {
+      $data = 'nodata';
+      echo json_encode($data);
    }
+   
 }
 
 ?>
